personal_finance_dashboard/requirements.txt
# FastAPI and ASGI server
fastapi==0.110.0
uvicorn==0.29.0

# Database (SQLAlchemy & PostgreSQL)
sqlalchemy==2.0.29
psycopg2-binary==2.9.9

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyjwt==2.8.0

# Environment Variables
python-dotenv==1.0.1

# Testing
pytest==8.1.1
httpx==0.27.0

# Alembic for database migrations
alembic==1.13.1
email-validator==1.1.3
